Reorganized dashboards karma tests.
